* logic is moved to abtract class
