+++ src/com/intellij/idea/plugin/hybris/toolwindow/RemoteHacConnectionDialog.kt
@@ -19,26 +19,17 @@
 package com.intellij.idea.plugin.hybris.toolwindow
 
 import com.intellij.credentialStore.Credentials
-import com.intellij.execution.wsl.WSLDistribution
-import com.intellij.execution.wsl.WslDistributionManager
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.settings.RemoteConnectionSettings
 import com.intellij.idea.plugin.hybris.tools.remote.RemoteConnectionScope
 import com.intellij.idea.plugin.hybris.tools.remote.http.HybrisHacHttpClient
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.ComboBox
-import com.intellij.openapi.util.registry.Registry
 import com.intellij.ui.EnumComboBoxModel
 import com.intellij.ui.SimpleListCellRenderer
-import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.dsl.builder.*
 import com.intellij.ui.layout.selected
 import java.awt.Component
-import java.util.*
-import javax.swing.DefaultComboBoxModel
-import javax.swing.JComboBox
-import javax.swing.JEditorPane
-import javax.swing.JLabel
 
 class RemoteHacConnectionDialog(
     project: Project,
@@ -47,9 +38,6 @@ class RemoteHacConnectionDialog(
 ) : AbstractRemoteConnectionDialog(project, parentComponent, settings, "Remote SAP Commerce Instance") {
 
     private lateinit var sslProtocolComboBox: ComboBox<String>
-    private lateinit var wslProxyCheckBox: JBCheckBox
-    private lateinit var wslProxyWarningComment: JEditorPane
-    private lateinit var wslDistributionText: Cell<JLabel>
 
 
     override fun createTestSettings() = with(RemoteConnectionSettings()) {
@@ -159,59 +147,7 @@ class RemoteHacConnectionDialog(
                     .component
             }.layout(RowLayout.PARENT_GRID)
             if (isWindows()) {
-                val distributions = WslDistributionManager.getInstance().installedDistributions
-                row {
-                    isWslCheckBox = checkBox("WSL")
-                        .bindSelected(settings::isWsl)
-                        .selected(false)
-                        .visible(distributions.isNotEmpty())
-                        .onChanged {
-                            val selected = isWslCheckBox.isSelected
-                            val multipleDistros = distributions.isNotEmpty()
-                            wslDistributionComboBox.isVisible = selected && multipleDistros
-                            wslDistributionText.visible(selected)
-                            wslProxyCheckBox.isVisible = selected
-                            wslProxyWarningComment.isVisible = selected
-                            urlPreviewLabel.text = generateUrl()
-                        }
-                        .component
-                }.layout(RowLayout.PARENT_GRID)
-                val installedDistros = distributions.map { it.msId }
-                if (installedDistros.isNotEmpty()) {
-                    row {
-                        wslDistributionText = label("WSL distribution:").visible(false)
-                        wslDistributionComboBox = comboBox(DefaultComboBoxModel(installedDistros.toTypedArray()))
-                            .align(AlignX.FILL)
-                            .visible(false)
-                            .onChanged {
-                                updateWslIp(distributions)
-                            }
-                            .component
-                    }.layout(RowLayout.PARENT_GRID)
-                } else {
-                    row {
-                        comment("No WSL distributions are installed.")
-                            .visible(false)
-                            .component
-                    }.layout(RowLayout.PARENT_GRID)
-                }
-                row {
-                    wslProxyCheckBox = checkBox("Enable wsl.proxy.connect.localhost")
-                        .comment("This will use the wsl.proxy.connect.localhost registry setting if available.")
-                        .visible(false)
-                        .selected(Registry.`is`(WSL_PROXY_CONNECT_LOCALHOST))
-                        .onChanged {
-                            Registry.run { get(WSL_PROXY_CONNECT_LOCALHOST).setValue(!`is`(WSL_PROXY_CONNECT_LOCALHOST)) }
-                            updateWslIp(distributions)
-                        }
-                        .component
-                }.layout(RowLayout.PARENT_GRID)
-                row {
-                    wslProxyWarningComment =
-                        comment("<strong>Warning:</strong> Connect to 127.0.0.1 on WSLProxy instead of public WSL IP which might be inaccessible due to routing issues.")
-                            .visible(false)
-                            .component
-                }.layout(RowLayout.PARENT_GRID)
+                wslHostConfiguration()
             }
         }
